requirements.txt
absl-py==1.4.0
certifi==2023.7.22
click==8.1.6
Flask
Flask-Cors
gast==0.4.0
google-pasta==0.2.0
grpcio==1.56.2
gunicorn
h5py==3.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Keras==2.12.0
Keras-Applications
Keras-Preprocessing
Markdown==3.4.4
MarkupSafe==2.1.3
numpy
Pillow==9.4.0
protobuf==3.20.3
PyYAML==6.0.1
scipy==1.10.1
six==1.16.0
tensorboard==2.12.3
tensorflow==2.12.0
tensorflow-estimator==2.12.0
termcolor==2.3.0
Werkzeug
wrapt==1.14.1
